requirements.txt
qrcode~=7.4.2
Django~=5.1
psycopg2-binary
djangorestframework
postgres
pycryptodome
djangorestframework-simplejwt